hdt-dashboard/requirements.txt
eel
azure-digitaltwins-core
azure-identity
python-dotenv
openai